requirements.txt
Django~=5.2
requests>=2.31.0
Pillow>=10.0.0
django-tinymce>=3.5.0
django-admin-sortable2>=2.1.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.7  # если используете PostgreSQL